Commands -> CommandRenderer (and move to own file).
